AnalyticsAI/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
duckdb
openai
python-dotenv
requests